wgpu: Move quad into its own structure under Descriptors
